% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gcbias.R
\name{add_gcbias}
\alias{add_gcbias}
\title{Calculate GC content of every DNA sequence}
\usage{
add_gcbias(object, genome = GenomeInfoDb::genome(object))
}
\arguments{
\item{object}{A SummarizedExperiment file}

\item{genome}{The genome the accessibility data was generated from (e.g. mm10, hg19).}
}
\description{
Calculates GC content of every DNA sequence from a SummarizedExperiment file.
}
\examples{
add_gcbias()
}
