% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addMatMat}
\alias{addMatMat}
\title{Add matrix through a matrix}
\usage{
addMatMat(A, B)
}
\arguments{
\item{A}{- matrix ( explain what this is)}

\item{B}{- vector ( explain what this is)}
}
\description{
Add matrix through a matrix
}
