% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addMatVec}
\alias{addMatVec}
\title{Add vector through a matrix}
\usage{
addMatVec(A, b)
}
\arguments{
\item{A}{- matrix ( explain what this is)}

\item{b}{- vector ( explain what this is)}
}
\description{
Add vector through a matrix
}
