% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-update.R
\name{annotate_probes_orgdb}
\alias{annotate_probes_orgdb}
\title{Annotate Probes Using Bioconductor OrgDb Packages}
\usage{
annotate_probes_orgdb(probe_ranges, txdb, orgdb)
}
\arguments{
\item{probe_ranges}{GRanges object with probe locations}

\item{txdb}{TxDb annotation package (e.g., TxDb.Hsapiens.UCSC.hg38.knownGene)}

\item{orgdb}{OrgDb annotation package (e.g., org.Hs.eg.db)}
}
\value{
Data frame with gene annotations
}
\description{
Alternative to biomaRt using local annotation packages.
Faster but requires package installation.
}
\examples{
\dontrun{
  library(TxDb.Hsapiens.UCSC.hg38.knownGene)
  library(org.Hs.eg.db)

  annotations <- annotate_probes_orgdb(
    probe_ranges,
    TxDb.Hsapiens.UCSC.hg38.knownGene,
    org.Hs.eg.db
  )
}
}
