% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{batch_process}
\alias{batch_process}
\title{Batch Process with Memory Management}
\usage{
batch_process(data, batch_size, fun, ...)
}
\arguments{
\item{data}{Data to process}

\item{batch_size}{Size of each batch}

\item{fun}{Function to apply to each batch}

\item{...}{Additional arguments to fun}
}
\value{
Combined results
}
\description{
Processes data in batches to manage memory
}
