% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedgraphs_to_accessibility_matrix.R
\name{bedgraphs_to_accessibility_matrix}
\alias{bedgraphs_to_accessibility_matrix}
\title{Bedgraph files to accessibility matrix}
\usage{
bedgraphs_to_accessibility_matrix(directory_bedgraph,
  column_accessibility = 4)
}
\arguments{
\item{directory_bedgraph}{Directory containing a list of bedgraph files. First 3 columns correspond to coordinates. User can specify which column contains accessibility values (automatically column 4).}

\item{column_accessibility}{The column containing the number of reads. Default is column 4.}
}
\description{
Creates an accessibility matrix from a list of bedgraph files. First 3 columns are coordinates, remaining columns correspond to cells, and every row is a DNA sequence.
}
\examples{
bedgraphs_to_accessibility_matrix()
}
