% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variation.R
\name{calculate_variation}
\alias{calculate_variation}
\title{Calculate cell-to-cell variation in chromatin accessibility}
\usage{
calculate_variation(scATACseq_experiment, ChIPseq_experiment)
}
\arguments{
\item{scATACseq_experiment}{A SummarizedExperiment containing scATACseq data.}

\item{ChIPseq_experiment}{A matrix containing ChIPseq data for matching DNA sequences.}
}
\description{
Calculates unnormalized cell-to-cell variation in chromatin accessibility using angular cosine distance inputted into principal coordinate analysis. Then each cell's distance from the centroid comprises the accessibility variation. Technical biases associated with GC content and mean accessibility are controlled for.
}
\examples{
calculate_variation()
}
