% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{callHotspots}
\alias{callHotspots}
\title{Call hotspots from rate files}
\usage{
callHotspots(
  prefix,
  suffix,
  chrList,
  minRate = 10,
  minFrac = 0.8,
  minLength = 1000,
  maxLength = 8000,
  maxTrim = NA,
  hotspotSmoothingLength = 2000,
  positionColumn = "Position.bp.",
  rateColumn = "Rate.cM.Mb.",
  verbose = 1,
  nCores = 1
)
}
\arguments{
\item{prefix}{for rate files, prefix, ie path to file without chromosome}

\item{suffix}{for rate files, suffix, ie path after chromosome}

\item{chrList}{chromosome list to find hotspots on}

\item{minRate}{minimum rate to declare hotspot}

\item{minFrac}{when a hotspot has been called, the continuous region with rate at this fraction of the maximum rate}

\item{minLength}{remove identified hotspots with a width smaller than this}

\item{maxLength}{remove hotspots with width smaller than this}

\item{hotspotSmoothingLength}{smooth recombination rates over this interval}

\item{positionColumn}{column name for physical position in recombination rate file}

\item{rateColumn}{column name for local (i.e. not cumulative) recombination rate in recombination rate file}

\item{verbose}{whether to print messages. 1 = Yes, 0 = No}

\item{nCores}{number of cores to use}
}
\description{
Call hotspots from rate files
}
