% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_distance.R
\name{centroid_distance}
\alias{centroid_distance}
\title{Distance from Centroid in Principal Coordinate Space using Angular Cosine Distance}
\usage{
centroid_distance(data, distance = "cosine", method = "centroid")
}
\arguments{
\item{data}{An accessibility matrix, with columns as cells and rows as DNA sequences.}
}
\description{
When using default parameters, calculates cell-to-cell variation (between columns) in accessibility using angular cosine distance inputted into principal coordinate analysis. The centroid of the cells in principal coordinate space is found, then each cell's distance from the centroid constitutes the variability.
}
\examples{
centroid_distance()
}
