% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{check_async_status}
\alias{check_async_status}
\title{Check the progress of async call}
\usage{
check_async_status()
}
\arguments{
\item{input}{required custom parameters for current model}
}
\value{
0 for sucess and 1 for error
}
\description{
Check the progress of async call
}
