% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_corrupt.R
\name{check_corrupt}
\alias{check_corrupt}
\title{Check the integrity of downloaded PRISM data}
\usage{
check_corrupt(type, minDate = NULL, maxDate = NULL, dates = NULL)
}
\arguments{
\item{type}{The type of data to download. Must be "ppt", "tmean", "tmin",
"tmax", "tdmean", "vpdmin", or "vpdmax". Note that \code{tmean ==  mean(tmin, tmax)}.}

\item{minDate}{Date to start downloading daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{maxDate}{Date to end downloading daily data.  Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{dates}{A vector of dates to download daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}
}
\value{
\code{logical} indicating whether the process succeeded.
}
\description{
Uses the \code{raster::stack} function to determine if the bil files are readable.
Any that are not readable are redownloaded.
}
