% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_critical_dependencies}
\alias{check_critical_dependencies}
\title{Check Critical Dependencies}
\usage{
check_critical_dependencies()
}
\description{
Verifies that critical Bioconductor packages are available
}
\keyword{internal}
