% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{check_packages}
\alias{check_packages}
\title{Check Package Installation}
\usage{
check_packages(packages, bioc = FALSE)
}
\arguments{
\item{packages}{Character vector of package names}

\item{bioc}{Logical indicating if packages are from Bioconductor}
}
\value{
Logical vector indicating which packages are installed
}
\description{
Checks if required packages are installed
}
