% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{check_platform_compatibility}
\alias{check_platform_compatibility}
\title{Check Platform Compatibility}
\usage{
check_platform_compatibility(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
Logical indicating compatibility
}
\description{
Checks if platform is compatible with PRISM workflow
}
