% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_prism_installation}
\alias{check_prism_installation}
\title{Check PRISM Installation}
\usage{
check_prism_installation()
}
\value{
Logical indicating if installation is complete
}
\description{
Verifies that PRISM is properly installed and configured
}
