% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{clean_gene_names}
\alias{clean_gene_names}
\title{Clean Gene Names}
\usage{
clean_gene_names(gene_names, to_upper = TRUE)
}
\arguments{
\item{gene_names}{Vector of gene names}

\item{to_upper}{Convert to uppercase (default: TRUE)}
}
\value{
Cleaned gene names
}
\description{
Standardizes gene names by removing special characters
}
\examples{
clean_gene_names(c("tp53", "BRCA-1", "egfr "))
}
