% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{convert_probe_ids}
\alias{convert_probe_ids}
\title{Convert Between Platform Formats}
\usage{
convert_probe_ids(probe_ids, from_platform, to_platform)
}
\arguments{
\item{probe_ids}{Vector of probe IDs}

\item{from_platform}{Source platform}

\item{to_platform}{Target platform}
}
\value{
Vector of converted probe IDs
}
\description{
Converts probe IDs between different platform formats
}
