% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_distance.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Angular cosine distance}
\usage{
cosine_distance(data)
}
\arguments{
\item{data}{An accessibility matrix, with columns as cells and rows as DNA sequences.}
}
\description{
Calculates angular cosine distance between columns (cells) of an accessibility matrix.
}
\examples{
cosine_distance()
}
