% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{count_unique}
\alias{count_unique}
\title{Count Unique Values}
\usage{
count_unique(data)
}
\arguments{
\item{data}{Data frame or matrix}
}
\value{
Named vector of unique counts
}
\description{
Counts unique values in each column
}
