% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Progress Bar Wrapper}
\usage{
create_progress_bar(n, title = "Progress")
}
\arguments{
\item{n}{Total number of iterations}

\item{title}{Progress bar title}
}
\value{
Progress bar object
}
\description{
Creates a simple progress bar
}
\keyword{internal}
