% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{cv}
\alias{cv}
\title{Calculate Coefficient of Variation}
\usage{
cv(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Remove NA values (default: TRUE)}
}
\value{
Coefficient of variation
}
\description{
Calculates CV (sd/mean)
}
