% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_early_prov.R
\name{del_early_prov}
\alias{del_early_prov}
\title{Delete early / provisional PRISM data if replaced by provisional / stable data.}
\usage{
del_early_prov(type, minDate = NULL, maxDate = NULL, dates = NULL)
}
\arguments{
\item{type}{The type of data to download. Must be "ppt", "tmean", "tmin",
"tmax", "tdmean", "vpdmin", or "vpdmax". Note that \code{tmean == mean(tmin, tmax)}.}

\item{minDate}{Date to start downloading daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{maxDate}{Date to end downloading daily data.  Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{dates}{A vector of dates to download daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}
}
\description{
Searches the download folder for duplicated PRISM data
and keeps only the newest version.
}
