% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{detect_from_gpl}
\alias{detect_from_gpl}
\title{Detect Platform from GPL ID}
\usage{
detect_from_gpl(gpl_id)
}
\arguments{
\item{gpl_id}{GEO platform ID}
}
\value{
Platform name or NULL
}
\description{
Detect Platform from GPL ID
}
\keyword{internal}
