% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{detect_from_probe_ids}
\alias{detect_from_probe_ids}
\title{Detect Platform from Probe IDs}
\usage{
detect_from_probe_ids(probe_ids)
}
\arguments{
\item{probe_ids}{Vector of probe IDs}
}
\value{
Platform name or NULL
}
\description{
Detect Platform from Probe IDs
}
\keyword{internal}
