% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect Outliers}
\usage{
detect_outliers(x, k = 1.5)
}
\arguments{
\item{x}{Numeric vector}

\item{k}{IQR multiplier (default: 1.5)}
}
\value{
Logical vector indicating outliers
}
\description{
Identifies outliers using IQR method
}
