% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert Data Frame to Matrix}
\usage{
df_to_matrix(df, row_names = 1)
}
\arguments{
\item{df}{Data frame to convert}

\item{row_names}{Column to use as row names (default: 1)}
}
\value{
Numeric matrix
}
\description{
Safely converts data frame to numeric matrix
}
