% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffplot.R
\name{diffplot}
\alias{diffplot}
\title{Partially Recursively Induced Structured Moderation}
\usage{
diffplot(l, dat, hvar, dot.size = 1, legend.size = 0.75)
}
\arguments{
\item{l}{An rpart output used to plot prism or hprism tree}

\item{dat}{The same data frame used to generate the prism or hprism object.}

\item{hvar}{The hierarchical level variable for hprism tree, which is specified within double quotes. If unspecified, prism tree is fitted instead. See Example for detail.}

\item{dot.size}{A numerical value giving the amount by which plotting should be magnified relative to the default.}

\item{legend.size}{A numerical value giving the amount by which text should be magnified relative to the default.}
}
\description{
Plot the HPRSIM hierarchical moderation plots for hprism object.
}
\details{
More can be learned by a deeper examination of the HPRISM tree topologies. Each branch path to a terminal node for an HPRISM tree explicitly details a discovered hierarchical interaction. This function plots the differences in (log) survival time between levels of the a focus variablefactor \emph{x} versus values of a contextual social determinant variable \emph{w}. Each dotted line represents a different terminal node from the HPRISM tree. Conditioning on a particular value of the social determinant variable, we see that if the data points on the line do not coincide, this indicates the first layer of interaction (i.e. the interaction of individual level variables defining the terminal nodes with the focus variable). If the lines are not parallel across different values of \emph{w}, this is indicative of the hierarchical interaction.
}
\examples{
\dontrun{
diffplot(l.hprism,dat=Leuk,hvar="tpi",dot.size=0.5)
diffplot(l.hprism1,dat=Leuk,hvar="tpi.m",dot.size=0.5)
}
}
