% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.is_retriable_draw_error}
\alias{.is_retriable_draw_error}
\title{Is this cov_bounds()/relative-covariance error a retriable degenerate draw?}
\usage{
.is_retriable_draw_error(message)
}
\description{
Is this cov_bounds()/relative-covariance error a retriable degenerate draw?
}
\keyword{internal}
