% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{.offdiag_pairs}
\alias{.offdiag_pairs}
\title{Off-diagonal (i < j) pair indices for a D x D matrix, column-major order}
\usage{
.offdiag_pairs(D)
}
\description{
Off-diagonal (i < j) pair indices for a D x D matrix, column-major order
}
\keyword{internal}
