% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.resolve_bootstrap_mode}
\alias{.resolve_bootstrap_mode}
\title{Normalize the bootstrap mode argument}
\usage{
.resolve_bootstrap_mode(bootstrap)
}
\description{
Normalize the bootstrap mode argument
}
\keyword{internal}
