% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.restore_rng_state}
\alias{.restore_rng_state}
\title{Restore the global RNG state saved by `.save_rng_state()`}
\usage{
.restore_rng_state(old)
}
\description{
Restore the global RNG state saved by `.save_rng_state()`
}
\keyword{internal}
