% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{.stream_open}
\alias{.stream_open}
\title{Open temporary binary files for streamed lower/upper/rel draw storage}
\usage{
.stream_open()
}
\description{
Open temporary binary files for streamed lower/upper/rel draw storage
}
\keyword{internal}
