% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{eset_to_matrix}
\alias{eset_to_matrix}
\title{Convert Expression Set to Matrix}
\usage{
eset_to_matrix(eset)
}
\arguments{
\item{eset}{ExpressionSet or SummarizedExperiment object}
}
\value{
Expression matrix
}
\description{
Extracts expression matrix from ExpressionSet or SummarizedExperiment
}
\examples{
\dontrun{
  expr_matrix <- eset_to_matrix(my_eset)
}
}
