% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{estimate_covariance}
\alias{estimate_covariance}
\title{Run Analysis on All Pairwise Taxa}
\usage{
estimate_covariance(
  Y,
  alpha = rep(0, nrow(Y)),
  rhobound = 0.8,
  S = 1000,
  variance = FALSE,
  upperx = 1
)
}
\arguments{
\item{Y}{A matrix of data with observations in columns and variables in rows.}

\item{alpha}{A numeric vector of priors for the Dirichlet distribution. Defaults to a vector of zeros.}

\item{rhobound}{A numeric value specifying the bound for the \code{rho1} and \code{rho2} parameters. Defaults to 0.8.}

\item{S}{An integer specifying the number of bootstrap samples. Increase to reduce Monte Carlo error. Defaults to 1000.}

\item{variance}{A boolean specifying whether to include diagonals of variance-covariance matrices (the variance). Defaults to FALSE.}

\item{upperx}{A numeric value specifying the upper bound of the variance of the scale. Defaults to 1.0.}
}
\value{
A list of dataframes containing the results of the analysis including estimated 95% confidence intervals, minimum and maximum values for estimated covariance, and finite sample covariances. \code{final_results} contains the data intended for forest_plot() and \code{all_inner_results} contains the data intended for sigmaplot().
}
\description{
This function runs a bootstrapped analysis on the input data matrix \code{Y}.
}
\examples{
# Example usage (You could also use the simulation function provided to generate sample data):
set.seed(123)
Y <- matrix(rnorm(1000), nrow = 10)
results <- estimate_covariance(Y)
}
