% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{estimate_memory}
\alias{estimate_memory}
\title{Estimate Memory Usage}
\usage{
estimate_memory(n_probes, n_samples, bytes_per_value = 8)
}
\arguments{
\item{n_probes}{Number of probes/genes}

\item{n_samples}{Number of samples}

\item{bytes_per_value}{Bytes per numeric value (default: 8)}
}
\value{
Character string with estimated memory
}
\description{
Estimates memory required for expression matrix
}
\examples{
estimate_memory(50000, 100)
}
