% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_table.R
\name{evaluation_table}
\alias{evaluation_table}
\title{Out-of-sample prediction evaluation}
\usage{
evaluation_table(claim_data, prism_prediction)
}
\arguments{
\item{claim_data}{the output of load_claim_data().}

\item{prism_prediction}{the output of back_test().}
}
\description{
Out-of-sample prediction evaluation
}
