% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_graph.R
\name{Example_graph}
\alias{Example_graph}
\title{Graph results from the example data}
\usage{
Example_graph(
  ListofTraits,
  Trait,
  ParametersTable,
  ThreshSelectionPvalues = 5e-08/length(ListofTraits)
)
}
\arguments{
\item{ListofTraits}{All traits in the analysis (necessary to calculate the p-value threshold)}

\item{Trait}{The trait of interest.}

\item{ParametersTable}{The parameters table for each pair of traits.}

\item{ThreshSelectionPvalues}{P-value threshold to select PRISM top variants, if you do not want
the default Bonferroni correction.}
}
\value{
a ggplot of all genetic variants with GWAS p-value on the x-axis, PRISM p-value
on the y-axis, and colors according to pleiotropy.
}
\description{
`Example_graph()` returns a ggplot of the results obtained from the example.
}
