% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_network.R
\name{Example_network}
\alias{Example_network}
\title{Graph network from the example data}
\usage{
Example_network(
  ListofTraits,
  Variant,
  ParametersTable,
  ThreshSelectionPvalues = 5e-08/(length(ListofTraits) - 1)
)
}
\arguments{
\item{ListofTraits}{All traits in the analysis (necessary to calculate the p-value threshold)}

\item{Variant}{The variant of interest.}

\item{ParametersTable}{The parameters table for each pair of traits.}

\item{ThreshSelectionPvalues}{P-value threshold to select PRISM top variants, if you do not want
the default Bonferroni correction.}
}
\value{
a deconvoluted causal variant network
}
\description{
`Example_network()` returns a ggplot of the results obtained from the example.
}
