% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{export_platform_info}
\alias{export_platform_info}
\title{Export Platform Information}
\usage{
export_platform_info(platform_id, output_file, format = "txt")
}
\arguments{
\item{platform_id}{GPL platform ID}

\item{output_file}{Output file path}

\item{format}{Output format: "txt", "json", "yaml"}
}
\description{
Exports platform information to file
}
