% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{find_common}
\alias{find_common}
\title{Find Common Elements}
\usage{
find_common(...)
}
\arguments{
\item{...}{Vectors or list of vectors}
}
\value{
Vector of common elements
}
\description{
Finds elements common to all vectors in a list
}
\examples{
find_common(c("A", "B", "C"), c("B", "C", "D"), c("C", "D", "E"))
}
