% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{format_number}
\alias{format_number}
\title{Format Number with Commas}
\usage{
format_number(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of decimal places (default: 0)}
}
\value{
Formatted character vector
}
\description{
Adds thousand separators to numbers
}
\examples{
format_number(1234567.89, digits = 2)
}
