% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy String Match}
\usage{
fuzzy_match(x, table, max_distance = 2)
}
\arguments{
\item{x}{Character vector to match}

\item{table}{Character vector to match against}

\item{max_distance}{Maximum edit distance (default: 2)}
}
\value{
Data frame with matches
}
\description{
Finds approximate string matches
}
