% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_affymetrix_annotation}
\alias{get_affymetrix_annotation}
\title{Get Affymetrix Annotation Package}
\usage{
get_affymetrix_annotation(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
Annotation package name
}
\description{
Returns the appropriate annotation package for Affymetrix platform
}
