% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_background_peaks.R
\name{get_background_peaks}
\alias{get_background_peaks}
\title{Background peak selection}
\usage{
get_background_peaks(scATACseq_experiment, ChIPseq_experiment)
}
\arguments{
\item{scATACseq_experiment}{A SummarizedExperiment with scATACseq data.}

\item{ChIPseq_experiment}{A ChIPseq matrix.}
}
\description{
Selects background peaks to control for technical biases associated with GC content and mean accessibility. Background peaks are drawn from non-TF bound peaks and matched peak-for-peak for GC content, and matrix-for-matrix for mean accessibility.
}
\examples{
get_background_peaks()
}
