% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{get_example_data}
\alias{get_example_data}
\title{Get PRISM Example Data}
\usage{
get_example_data(dataset = NULL)
}
\arguments{
\item{dataset}{Name of example dataset}
}
\value{
Path to example data
}
\description{
Returns path to example data files
}
\examples{
\dontrun{
  example_file <- get_example_data("sample_probes")
}
}
