% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{get_file_size}
\alias{get_file_size}
\title{Check File Size}
\usage{
get_file_size(file)
}
\arguments{
\item{file}{Path to file}
}
\value{
Character string with file size
}
\description{
Returns human-readable file size
}
