% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-update.R
\name{get_gene_metadata}
\alias{get_gene_metadata}
\title{Retrieve Additional Gene Metadata}
\usage{
get_gene_metadata(
  gene_ids,
  species = "hsapiens",
  attributes = c("go_id", "name_1006", "namespace_1003")
)
}
\arguments{
\item{gene_ids}{Vector of Ensembl gene IDs}

\item{species}{Species name (default: "hsapiens")}

\item{attributes}{Additional biomaRt attributes to retrieve}
}
\value{
Data frame with gene metadata
}
\description{
Fetches extra information like GO terms, pathways, etc.
}
\examples{
\dontrun{
  metadata <- get_gene_metadata(
    c("ENSG00000139618", "ENSG00000141510"),
    species = "hsapiens"
  )
}
}
