% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_illumina_annotation}
\alias{get_illumina_annotation}
\title{Get Illumina Annotation Package}
\usage{
get_illumina_annotation(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
Annotation package name
}
\description{
Returns the appropriate annotation package for Illumina platform
}
