% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_metadata}
\alias{get_metadata}
\title{Get PRISM metadata}
\usage{
get_metadata(type, minDate, maxDate)
}
\arguments{
\item{type}{The type of data to download, must be "ppt", "tmean", "tmin", or "tmax".
Note that tmean == mean(tmin, tmax).}

\item{minDate}{a valid iso-8601 (e.g. YYYY-MM-DD) date to start to download data.}

\item{maxDate}{a valid iso-8601 (e.g. YYYY-MM-DD) date to end downloading data.}
}
\value{
list of data.frames containing metadata. If only
one date is requested, the function returns the data.frame.
}
\description{
Retrieves PRISM metadata for a given type and
date range. The information is retrieved from the .info.txt file.
}

