% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_platform_annotation}
\alias{get_platform_annotation}
\title{Get Platform Annotation Package}
\usage{
get_platform_annotation(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
Annotation package name
}
\description{
Returns appropriate annotation package for any platform
}
