% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_platform_genome}
\alias{get_platform_genome}
\title{Get Platform Genome Build}
\usage{
get_platform_genome(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
Character string with genome build
}
\description{
Returns the recommended genome build for a platform
}
