% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{get_platform_specs}
\alias{get_platform_specs}
\title{Get Platform Specifications}
\usage{
get_platform_specs(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
List with platform specifications
}
\description{
Returns detailed specifications for a platform
}
