% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prism_annual.R
\name{get_prism_annual}
\alias{get_prism_annual}
\title{Download annual daily averages}
\usage{
get_prism_annual(type, years = NULL, keepZip = TRUE,
  keep_pre81_months = FALSE)
}
\arguments{
\item{type}{The type of data to download, must be "ppt", "tmean", "tmin",
"tmax", "tdmean", "vpdmin", or "vpdmax". Note that \code{tmean == mean(tmin, tmax)}.}

\item{years}{a valid numeric year, or vector of years, to download data for.
If no month is specified, year averages for that year will be downloaded.}

\item{keepZip}{if \code{TRUE}, leave the downloaded zip files in your
'prism.path', if \code{FALSE}, they will be deleted.}

\item{keep_pre81_months}{The pre-1981 data includes all monthly data and the
annual data for the specified year. If you need annual and monthly data it
is advantageous to keep all the monthly data when downloading the annual
data so you don't have to download the zip file again. When downloading
annual data, this defaults to \code{FALSE}. When downloading monthly data, this
defaults to \code{TRUE}.}
}
\description{
Download annual daily average data from the prism project at 4km grid cell
resolution for precipitation, mean, min and max temperature
}
\details{
Data is available from 1891 until 2014, however you have to download
all data for years prior to 1981.  Therefore if you enter a vector of years
that bounds 1981, you will automatically download all data for all years in
the vector.  If the "all" parameter is set to TRUE, it will override any
months entered and download all data. Data will be downloaded for all months
in all the years in the vectors supplied. You must make sure that you have
set up a valid download directory.  This must be set as
\code{options(prism.path = "YOURPATH")}.
}
\examples{
\dontrun{
### Get all the data for January from 1990 to 2000
get_prism_annual(type="tmean", year = 1990:2000, keepZip=FALSE)
}

}
