% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_prism_option}
\alias{get_prism_option}
\title{Get PRISM Options}
\usage{
get_prism_option(option = NULL)
}
\arguments{
\item{option}{Option name (if NULL, returns all PRISM options)}
}
\value{
Option value or list of all options
}
\description{
Retrieves PRISM package options
}
\examples{
get_prism_option("verbose")
get_prism_option()  # Get all options
}
