% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prism_pd.R
\name{get_prism_pd}
\alias{get_prism_pd}
\title{\code{get_prism_pd()} downloads PRISM data based on the folder names (\code{pd}) that
are created after downloading data.}
\usage{
get_prism_pd(
  pd,
  keepZip = prism_get_keepZip(),
  service = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{pd}{Character vector. Should be in the form
\code{"prism_[type]_[region]_[resolution]_[date]{_avg_30y}"}.}
}
\description{
\code{get_prism_pd()} downloads PRISM data based on the folder names (\code{pd}) that
are created after downloading data.
}
\examples{
\dontrun{
get_prism_pd("prism_ppt_us_25m_199107")
get_prism_pd(c("prism_ppt_us_25m_199107", "prism_ppt_us_25m_199108"))

# redownload out of date data
pd <- pd_check_versions()
get_prism_pd(pd, overwrite = TRUE)
}

#' @rdname get_prism_data

}
