% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{get_prism_session_info}
\alias{get_prism_session_info}
\title{Get PRISM Session Info}
\usage{
get_prism_session_info()
}
\value{
List with session details
}
\description{
Returns detailed session information for reproducibility
}
