% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_recent_filenames}
\alias{get_recent_filenames}
\title{Get daily data file names for last 2 years}
\usage{
get_recent_filenames(type, frequency)
}
\arguments{
\item{type}{The type of data to download, must be "ppt", "tmean", "tmin", "tmax", or "all",
which downloads "ppt", "tmin", and "tmax". Note that tmean == mean(tmin, tmax).}

\item{frequency}{\code{character} for the frequency. One of
"daily" or "monthly".}
}
\description{
Utility function to download all filenames
for the last two years. This is to determine which are
currently showing as provisional, early, or stable.
}

