% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-remap.R
\name{get_remapping_summary}
\alias{get_remapping_summary}
\title{Get Remapping Summary}
\usage{
get_remapping_summary(probe_expr, gene_expr, probe_gene_map)
}
\arguments{
\item{probe_expr}{Original probe expression matrix}

\item{gene_expr}{Remapped gene expression matrix}

\item{probe_gene_map}{Probe-gene mapping}
}
\value{
List with summary statistics
}
\description{
Summarizes the remapping process
}
