% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getEnrichmentFromSequences}
\alias{getEnrichmentFromSequences}
\title{Get enrichment for a set of hotspots}
\usage{
getEnrichmentFromSequences(
  seqsH,
  seqsC,
  hotspotCenterDist = 200,
  K = 8,
  verbose = 1,
  nCores = 1
)
}
\arguments{
\item{seqsH}{hotspot sequences}

\item{seqsC}{coldspot sequences}

\item{K}{k-mer length to seed over}

\item{verbose}{whether to print messages. 2 = frequent, 1 = moderate, 0 = No}

\item{nCores}{How many computer cores to use}

\item{hotCenterDist}{Distance from hotspot centre each way to look for enrichment}

\item{filterSequences}{Whether to filter hotspots for repeats and simple repeats}

\item{rmsk}{Path to repeat masker file}

\item{simpleRepeats}{Path to simple repeats file}
}
\description{
Get enrichment for a set of hotspots
}
