% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotifs.R
\name{getMotifs}
\alias{getMotifs}
\title{Refine motif in hotspots from seed sequence}
\usage{
getMotifs(
  hot,
  seqsH,
  seqsC,
  hotExclude = NULL,
  eligible = NULL,
  nCores = 1,
  K = 8,
  hotspotCenterDist = 200,
  filterSequences = FALSE,
  fasta = NULL,
  rmsk = NULL,
  simpleRepeats = NULL,
  discovery = TRUE,
  initializationMethod = "advanced",
  dataType = "normal",
  pwmValue = 0.9,
  alpha = 0.5,
  incprob = 0.99999,
  maxits = 30,
  plen = 0.05,
  updatemot = 1,
  updatealpha = 1,
  ourprior = NULL,
  updateprior = 1,
  bg = -1,
  plotting = "pdf",
  plotPrefix = "getMotifs.plot",
  verbose = 1,
  scorematset = NULL,
  dimvec = NULL,
  pwm = NULL
)
}
\arguments{
\item{hot}{A set of hotspots}

\item{seqsH}{A list of hotspots (use initializationMethod advanced, dataType sequence)}

\item{seqsC}{A list of coldspots (use initializationMethod advanced, dataType sequence)}

\item{hotExclude}{Hotspots to exclude (for instance, if these are known to contain matches to another motif)}

\item{eligible}{k-mers already used as seeds (or NULL)}

\item{nCores}{Number of computer cores to use}

\item{K}{Length of k-mers to consider for seeding}

\item{fasta}{Path to fasta file}

\item{rmsk}{Path to repeat masker file (assumes preparation already performed)}

\item{simpleRepeats}{Path to simple repeats file (assumes preparation already performed)}

\item{initializationMethod}{Whether to initialize using a single top k-mer (simple), by constructing a rudimentary PWM (advanced), or to use pre-existing motif probabilities (existing)}

\item{dataType}{Whether input data is sequences (sequence) or chromosome and fasta based}

\item{pwmValue}{If initializing simply, the probability the PWM contains the sequence of the k-mer}

\item{alpha}{To complete}

\item{incprob}{To complete}

\item{maxits}{To complete}

\item{plen}{To complete}

\item{updatemot}{To complete}

\item{updatealpha}{To complete}

\item{ourprior}{To complete}

\item{updateprior}{To complete}

\item{bg}{To complete}

\item{plotting}{Whether to print an output PDF ("pdf"), to screen ("screen"), or not (anything else)}

\item{plotPrefix}{If outputting PDFs, what to call them}

\item{verbose}{Verbosity. 1 = Standard, 2 = Detailed, otherwise no printing}

\item{scorematset}{To complete}

\item{dimvec}{To complete}

\item{pwm}{When using your own initialization,}
}
\description{
Refine motif in hotspots from seed sequence
}
