% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSequences}
\alias{getSequences}
\title{Get sequences for a set of hotspots}
\usage{
getSequences(
  chrList,
  hot,
  nCores = 1,
  filterSequences = TRUE,
  rmsk = NULL,
  simpleRepeats = NULL,
  verbose = 1
)
}
\arguments{
\item{chrList}{List of chromosomes to run over}

\item{hot}{A set of hotspots}

\item{nCores}{How many computer cores to use}

\item{filterSequences}{Whether to filter hotspots for repeats and simple repeats}

\item{rmsk}{Path to repeat masker file}

\item{simpleRepeats}{Path to simple repeats file}

\item{verbose}{whether to print messages. 2 = frequent, 1 = moderate, 0 = No}
}
\description{
Get sequences for a set of hotspots
}
