% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In Operator}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{Vector of values}

\item{y}{Vector to check against}
}
\value{
Logical vector
}
\description{
Negation of \%in\%
}
\keyword{internal}
