% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (returned if x is NULL)}
}
\value{
x if not NULL, otherwise y
}
\description{
Returns the right-hand side if left-hand side is NULL
}
\keyword{internal}
