% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-geo.R
\name{import_geo_data}
\alias{import_geo_data}
\title{Import Microarray Data from GEO}
\usage{
import_geo_data(geo_id, platform = NULL, destdir = tempdir())
}
\arguments{
\item{geo_id}{GEO accession number (e.g., "GSE12345")}

\item{platform}{Platform GPL ID (optional, auto-detected if NULL)}

\item{destdir}{Directory to store downloaded files}
}
\value{
A prism_array object
}
\description{
Downloads and processes microarray data from NCBI GEO database
}
\examples{
\dontrun{
  geo_data <- import_geo_data("GSE12345")
}
}
