% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-local.R
\name{import_probe_fasta}
\alias{import_probe_fasta}
\title{Import Probe Sequences from FASTA}
\usage{
import_probe_fasta(file)
}
\arguments{
\item{file}{Path to FASTA file}
}
\value{
DNAStringSet with probe sequences
}
\description{
Reads probe sequences from FASTA file
}
\examples{
\dontrun{
  sequences <- import_probe_fasta("probes.fasta")
}
}
