% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{increment}
\alias{increment}
\title{For a vector y of length yT, return a (1-based) vector where entry i counts the number of times integer i-1 was seen in the data}
\usage{
increment(y, yT, xT)
}
\arguments{
\item{y}{vector of integers}

\item{yT}{length of y}

\item{xT}{maxmimum integer to count to}
}
\description{
For a vector y of length yT, return a (1-based) vector where entry i counts the number of times integer i-1 was seen in the data
}
