% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{init_log_file}
\alias{init_log_file}
\title{Create Log File}
\usage{
init_log_file(log_file, append = FALSE)
}
\arguments{
\item{log_file}{Path to log file}

\item{append}{Append to existing file (default: FALSE)}
}
\value{
Log file connection
}
\description{
Initializes a log file for tracking analysis
}
