% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initializePWM}
\alias{initializePWM}
\title{After calculating enrichment, initialize a PWM}
\usage{
initializePWM(
  motifEnrichment,
  K = 8,
  initializationMethod = "simple",
  eligible = NULL,
  verbose = 1,
  pwmValue = 0.9
)
}
\arguments{
\item{motifEnrichment}{Specific input from getEnrichment}

\item{K}{k-mer length to consider}

\item{initializationMethod}{Either "simple" or "advanced". Simple just takes top k-mer, advanced builds simple PWM}

\item{eligible}{Which k-mers are eligible to be considered}

\item{verbose}{Verbosity. 1 = Full}
}
\description{
After calculating enrichment, initialize a PWM
}
