% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{is_agilent_control}
\alias{is_agilent_control}
\title{Handle Agilent Control Probes}
\usage{
is_agilent_control(probe_ids, annotations = NULL)
}
\arguments{
\item{probe_ids}{Vector of probe IDs}

\item{annotations}{Optional probe annotation data frame}
}
\value{
Logical vector indicating control probes
}
\description{
Identifies and filters Agilent control probes
}
