% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{is_illumina_control}
\alias{is_illumina_control}
\title{Handle Illumina Control Probes}
\usage{
is_illumina_control(probe_ids)
}
\arguments{
\item{probe_ids}{Vector of probe IDs}
}
\value{
Logical vector indicating control probes
}
\description{
Identifies and filters Illumina control probes
}
