% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{list_supported_platforms}
\alias{list_supported_platforms}
\title{List Supported Platforms}
\usage{
list_supported_platforms()

list_supported_platforms()
}
\value{
Data frame with supported platforms

Data frame with supported platforms
}
\description{
Returns list of platforms with built-in support

Returns list of platforms with built-in support
}
