% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{list_visualization_functions}
\alias{list_visualization_functions}
\title{Get Available Visualization Functions}
\usage{
list_visualization_functions()
}
\value{
Character vector of function names
}
\description{
Lists all available plotting functions in PRISM
}
