% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_5y_search_data.R
\name{load_5y_search_data}
\alias{load_5y_search_data}
\title{Load Goolge Trends data and initial claims data}
\usage{
load_5y_search_data(folder = "0408")
}
\arguments{
\item{folder}{foldernames for a certain periods of Google Trends data. The foldernames are "0408", "0610", "0812", "1014", "1216". For example, the folder "0408" is for 2004-2008.}
}
\value{
A list of following named xts objects
\itemize{
\item \code{claim.data} unemployment initial claim data of the same span as Google Trend data.
\item \code{claim.all} load all unemployment initial claim data since 1967
\item \code{claim.early} unemployment initial claim data from 1980-01-06 to the start of \code{claim.data}.
\item \code{allSearch} Google Trends data of a span over five years. It is in the scale of 0 -- 100.
}
}
\description{
Load weekly unemployment initial claim data and related Google Trend data over 5-year span (each week ends on the Saturday). The list of Google search terms is the same as in paper.
}
