% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{load_affymetrix_sequences}
\alias{load_affymetrix_sequences}
\title{Load Affymetrix Probe Sequences}
\usage{
load_affymetrix_sequences(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
DNAStringSet with probe sequences
}
\description{
Loads probe sequences for Affymetrix platforms
}
\examples{
\dontrun{
  sequences <- load_affymetrix_sequences("GPL570")
}
}
