% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{load_agilent_sequences}
\alias{load_agilent_sequences}
\title{Load Agilent Probe Sequences}
\usage{
load_agilent_sequences(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
DNAStringSet with probe sequences
}
\description{
Loads probe sequences for Agilent platforms
}
