% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_claim_data.R
\name{load_claim_data}
\alias{load_claim_data}
\title{Load unemployment initial claims data}
\usage{
load_claim_data(GT.startDate = "2004-01-03", GT.endDate = "2016-12-31")
}
\arguments{
\item{GT.startDate}{start date of claim data}

\item{GT.endDate}{end date of claim data}
}
\value{
A list of following named xts objects
\itemize{
\item \code{claim.data} unemployment initial claim data from GT.startDate to GT.endDate.
\item \code{claim.all} load all unemployment initial claim data since 1967
\item \code{claim.early} unemployment initial claim data prior to GT.startDate
}
}
\description{
Load weekly unemployment initial claim data (each week ends on the Saturday).
}
