% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{load_custom_sequences}
\alias{load_custom_sequences}
\title{Load Custom Platform Sequences}
\usage{
load_custom_sequences(sequence_file, platform_id = NULL)
}
\arguments{
\item{sequence_file}{Path to FASTA file with probe sequences}

\item{platform_id}{Platform identifier}
}
\value{
DNAStringSet with probe sequences
}
\description{
Loads probe sequences from a custom file
}
