% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{load_illumina_sequences}
\alias{load_illumina_sequences}
\title{Load Illumina Probe Sequences}
\usage{
load_illumina_sequences(platform_id)
}
\arguments{
\item{platform_id}{GPL platform ID}
}
\value{
DNAStringSet with probe sequences
}
\description{
Loads probe sequences for Illumina platforms
}
