% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{log_message}
\alias{log_message}
\title{Log Message with Timestamp}
\usage{
log_message(..., level = "INFO")
}
\arguments{
\item{...}{Message components}

\item{level}{Log level: "INFO", "WARNING", "ERROR"}
}
\description{
Prints message with timestamp
}
