% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ls_prism_data}
\alias{ls_prism_data}
\title{List available datasets}
\usage{
ls_prism_data(absPath = F)
}
\arguments{
\item{absPath}{TRUE if you want to return the absolute path.}
}
\description{
List the available data sets to load that have already been downloaded.
}
\examples{
\dontrun{
ls_prism_data()
}
}

