% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{mad_score}
\alias{mad_score}
\title{Calculate Median Absolute Deviation}
\usage{
mad_score(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Remove NA values (default: TRUE)}
}
\value{
MAD value
}
\description{
Robust measure of variability
}
