% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_summarized_experiment.R
\name{matrix_to_summarized_experiment}
\alias{matrix_to_summarized_experiment}
\title{Creates a Summarized Experiment from an accessibility matrix}
\usage{
matrix_to_summarized_experiment(accessibility_matrix, genome)
}
\arguments{
\item{accessibility_matrix}{A matrix with first 3 columns containing coordinates. Remaining columns correspond to cells. Rows correspond to DNA sequences. Accessibility data should be binarized.}

\item{genome}{The genome the accessibility data was generated from (e.g. mm10, hg19).}
}
\description{
Converts an accessibility matrix into a SummarizedExperiment file.
}
\examples{
matrix_to_summarized_experiment()
}
