% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1client_lib.R
\name{model_run_async}
\alias{model_run_async}
\title{Executes PRISM model in async model}
\usage{
model_run_async(input = NULL)
}
\arguments{
\item{input}{required custom parameters for current model}
}
\value{
0 for sucess and 1 for error
}
\description{
Executes PRISM model in async model
}
