% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mon_to_string}
\alias{mon_to_string}
\title{helper function for handling months}
\usage{
mon_to_string(month)
}
\arguments{
\item{month}{a numeric vector of months (month must be > 0 and <= 12)}
}
\value{
a character vector (same length as \code{month}) with 2 char month strings.
}
\description{
Handle numeric month to string conversions
}
\examples{
\dontrun{
  mon_to_string(month = c(1, 3, 2))
  mon_to_string(month = 12)
}
}

