% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-classes.R
\name{new_prism_annotation}
\alias{new_prism_annotation}
\title{PRISM Annotation Results}
\usage{
new_prism_annotation(alignments, annotations, qc_flags, summary)
}
\arguments{
\item{alignments}{GRanges object with probe alignments}

\item{annotations}{Data frame with gene annotations}

\item{qc_flags}{Data frame with quality control flags}

\item{summary}{List with summary statistics}
}
\value{
A prism_annotation object
}
\description{
S3 class for storing re-annotation results
}
