% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairwise_pipeline.R
\name{Pairwise_pipeline}
\alias{Pairwise_pipeline}
\title{Pairwise pipeline of PRISM}
\usage{
Pairwise_pipeline(
  ListofTraits,
  ParametersTable,
  Index,
  sourceGWAS,
  NbCores,
  gzip,
  pU,
  Minimum_MAF
)
}
\arguments{
\item{ListofTraits}{The list of traits to be processed.}

\item{ParametersTable}{The parameters table for each pair of traits.}

\item{Index}{A dataframe with all variants and their LDscores.}

\item{sourceGWAS}{The folder where GWAS summary statistics files are held.}

\item{NbCores}{The number of cores to use.}

\item{gzip}{TRUE to gzip created files, FALSE otherwise.}

\item{pU}{Polygenicity of the confounder U.}

\item{Minimum_MAF}{Genetic variants with minimum allele frequency below this value will be filtered out.}
}
\description{
`Pairwise_pipeline()` handles the pairwise process of PRISM. It takes as input
the data and outputs in the Traitwise/ folder the likelihoods and scores of
each variant to belong to the different models, for each pair of traits.


All parameters are passed from `PRISM_main()` and are more detailed over there.
}
