% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{parse_affymetrix_ids}
\alias{parse_affymetrix_ids}
\title{Parse Affymetrix Probe IDs}
\usage{
parse_affymetrix_ids(probe_ids)
}
\arguments{
\item{probe_ids}{Vector of Affymetrix probe IDs}
}
\value{
Data frame with parsed information
}
\description{
Extracts information from Affymetrix probe IDs
}
\examples{
parse_affymetrix_ids(c("1007_s_at", "1053_at", "117_at"))
}
