% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{parse_agilent_ids}
\alias{parse_agilent_ids}
\title{Parse Agilent Probe IDs}
\usage{
parse_agilent_ids(probe_ids)
}
\arguments{
\item{probe_ids}{Vector of Agilent probe IDs}
}
\value{
Data frame with parsed information
}
\description{
Extracts information from Agilent probe IDs
}
\examples{
parse_agilent_ids(c("A_23_P123456", "A_24_P234567"))
}
