% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{parse_illumina_ids}
\alias{parse_illumina_ids}
\title{Parse Illumina Probe IDs}
\usage{
parse_illumina_ids(probe_ids)
}
\arguments{
\item{probe_ids}{Vector of Illumina probe IDs}
}
\value{
Data frame with parsed information
}
\description{
Extracts information from Illumina probe IDs
}
\examples{
parse_illumina_ids(c("ILMN_1234567", "ILMN_2345678"))
}
