% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_get_md.R
\name{pd_get_md}
\alias{pd_get_md}
\title{Get prism metadata}
\usage{
pd_get_md(pd)
}
\arguments{
\item{pd}{prism data character vector.}
}
\value{
data.frame containing metadata for all specified prism data.
}
\description{
Retrieves prism metadata from the specified prism data. "prism data", i.e.,
\code{pd} are the folder names returned by \code{\link[=prism_archive_ls]{prism_archive_ls()}} or
\code{\link[=prism_archive_subset]{prism_archive_subset()}}. A warning is provided if the specified prism data
do not exist in the archive.
}
\details{
The metadata includes a subset of the the following variables from the
.info.txt file for daily, monthly, and annual, time series and normals data:
\itemize{
\item PRISM_DATASET_FILENAME
\item PRISM_DATASET_CREATE_DATE
\item PRISM_DATASET_TYPE
\item PRISM_DATASET_VERSION
\item PRISM_CODE_VERSION
\item PRISM_DATASET_REMARKS
\item PRISM_SOURCE_FILENAME
\item PRISM_SOURCE_CREATE_DATE
\item PRISM_DATASET_RELEASE_NUMBER
}

Additionally, two local variables are added identifying where the file is
located on the local system:
\itemize{
\item file_path
\item folder_path
}

Not all metadata includes all variables, so only the variables that are
included in the specified \code{pd} are included in the returned data frame. If
not all variables are found in all of the \code{pd}, \code{NAs} are introduced into
the missing variables.
}
\examples{
\dontrun{
#' # Assumes 2000-2002 annual precipitation data is already downloaded
pd <- prism_archive_subset('ppt', 'annual', years = 2000:2002)
df <- pd_get_md(pd)
head(df)
}

}
