% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_alignment_stats}
\alias{plot_alignment_stats}
\title{Plot Alignment Statistics}
\usage{
plot_alignment_stats(alignments, max_locations = 10)
}
\arguments{
\item{alignments}{GRanges object from align_probes_to_genome}

\item{max_locations}{Maximum locations to show on x-axis (default: 10)}
}
\value{
ggplot object
}
\description{
Visualizes probe alignment distribution
}
\examples{
\dontrun{
  plot_alignment_stats(alignments)
}
}
