% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_annotation_coverage}
\alias{plot_annotation_coverage}
\title{Plot Annotation Coverage}
\usage{
plot_annotation_coverage(qc_results)
}
\arguments{
\item{qc_results}{Data frame from flag_problematic_probes}
}
\value{
ggplot object
}
\description{
Shows what percentage of probes have annotations
}
\examples{
\dontrun{
  plot_annotation_coverage(qc_results)
}
}
