% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_gene_biotypes}
\alias{plot_gene_biotypes}
\title{Plot Gene Biotype Distribution}
\usage{
plot_gene_biotypes(annotations, top_n = 10)
}
\arguments{
\item{annotations}{Data frame from update_gene_annotations}

\item{top_n}{Show top N biotypes (default: 10)}
}
\value{
ggplot object
}
\description{
Shows distribution of gene biotypes in annotations
}
\examples{
\dontrun{
  plot_gene_biotypes(annotations)
}
}
