% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA of Expression Data}
\usage{
plot_pca(expression_matrix, sample_groups = NULL, n_top = 500, pcs = c(1, 2))
}
\arguments{
\item{expression_matrix}{Expression matrix (genes x samples)}

\item{sample_groups}{Optional vector of sample group labels}

\item{n_top}{Number of top variable genes to use (default: 500)}

\item{pcs}{Principal components to plot (default: c(1, 2))}
}
\value{
ggplot object
}
\description{
Principal component analysis visualization
}
\examples{
\dontrun{
  plot_pca(gene_expr, sample_groups = c(rep("Control", 5), rep("Treatment", 5)))
}
}
