% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_probe_locations}
\alias{plot_probe_locations}
\title{Plot Probe Locations on Genome}
\usage{
plot_probe_locations(alignments, chromosome, start = NULL, end = NULL)
}
\arguments{
\item{alignments}{GRanges object from align_probes_to_genome}

\item{chromosome}{Chromosome to plot (e.g., "chr1")}

\item{start}{Start position (optional)}

\item{end}{End position (optional)}
}
\value{
ggplot object
}
\description{
Visualizes probe locations along a chromosome
}
\examples{
\dontrun{
  plot_probe_locations(alignments, "chr1", start = 1e6, end = 2e6)
}
}
