% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_qc_interactive}
\alias{plot_qc_interactive}
\title{Create Interactive QC Plot}
\usage{
plot_qc_interactive(qc_results)
}
\arguments{
\item{qc_results}{Data frame from flag_problematic_probes}
}
\value{
plotly object
}
\description{
Creates an interactive plotly visualization
}
\examples{
\dontrun{
  interactive_plot <- plot_qc_interactive(qc_results)
  interactive_plot
}
}
