% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_example_data.R
\name{Prepare_example_data}
\alias{Prepare_example_data}
\title{Prepare example data}
\usage{
Prepare_example_data(gzip = F)
}
\arguments{
\item{gzip}{TRUE if you want the example data to be gzipped, FALSE otherwise}
}
\description{
`Prepare_example_data()` creates a Data/ folder in the current working directory
and writes the dataset necessary to run the PleioVar example.
}
