% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepFasta}
\alias{prepFasta}
\title{Convert fasta into RData object files}
\usage{
prepFasta(fasta, chrList, verbose = 1, nCores = 1)
}
\arguments{
\item{fasta}{path to fasta file}

\item{chrList}{vector of chromosome names to prep}

\item{verbose}{whether to print messages. verbose 1 = Yes, 0 = No}

\item{nCores}{number of cores to use}
}
\description{
Convert fasta into RData object files
}
