% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{print.prism_cov_bounds}
\alias{print.prism_cov_bounds}
\title{Print covariance bounds}
\usage{
\method{print}{prism_cov_bounds}(x, ...)
}
\arguments{
\item{x}{A `prism_cov_bounds` object.}

\item{...}{Unused.}
}
\description{
Print covariance bounds
}
