% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-imports.R
\docType{package}
\name{prism-package}
\alias{prism}
\alias{prism-package}
\title{prism: Probe Re-Identification and Sequence Mapping}
\description{
Re-annotates legacy microarray datasets using current genome builds and gene annotations. Maps probe sequences to modern reference genomes, updates gene annotations from Ensembl, identifies problematic probes, and reprocesses expression matrices. Enables researchers to extract new insights from thousands of legacy datasets in repositories like GEO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markdunning/prism}
  \item Report bugs at \url{https://github.com/markdunning/prism/issues}
}

}
\author{
\strong{Maintainer}: Mark Dunning \email{m.dunning@imperial.ac.uk} (\href{https://orcid.org/0000-0002-8853-9435}{ORCID})

}
\keyword{internal}
