% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_archive_ls.R
\name{prism_archive_ls}
\alias{prism_archive_ls}
\title{List available prism data}
\usage{
prism_archive_ls()
}
\value{
\code{prism_archive_ls()} returns a character vector.
}
\description{
\code{prism_archive_ls()} lists all available prism data (all variables and all
temporal periods) that are available in the local archive, i.e., they
have already been downloaded and are available in \code{\link[=prism_get_dl_dir]{prism_get_dl_dir()}}.
\code{\link[=prism_archive_subset]{prism_archive_subset()}} can be used to subset the archive based on specified
variables and temporal periods.
}
\examples{
\dontrun{
# Get prism data names, used in many other prism* functions 
get_prism_dailys(
  type="tmean", 
  minDate = "2013-06-01", 
  maxDate = "2013-06-14", 
  keepZip = FALSE
)
prism_archive_ls()
}

}
\seealso{
\code{\link[=prism_archive_subset]{prism_archive_subset()}}
}
