% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prism_check}
\alias{prism_check}
\title{Helper function to check if files already exist}
\usage{
prism_check(prismfile)
}
\arguments{
\item{prismfile}{a list of full paths for prism files}
}
\value{
a character vector of file names that already exist
}
\description{
check if files exist
}

