% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_data_get_md.R
\name{prism_data_get_md}
\alias{prism_data_get_md}
\title{Get prism metadata}
\usage{
prism_data_get_md(
  type,
  temp_period,
  years = NULL,
  mon = NULL,
  minDate = NULL,
  maxDate = NULL,
  dates = NULL,
  resolution = NULL
)
}
\arguments{
\item{type}{The type of data you want to subset. Must be "ppt", "tmean",
"tmin", "tmax", "tdmean", "vpdmin", or "vpdmax".}

\item{temp_period}{The temporal period to subset. Must be "annual",
"monthly", "daily", "monthly normals", or "annual normals".}

\item{years}{Valid numeric year, or vector of years.}

\item{mon}{Valid numeric month, or vector of months.}

\item{minDate}{Date to start subsetting daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{maxDate}{Date to end subsetting daily data.  Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{dates}{A vector of daily dates to subset. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{resolution}{The spatial resolution of the data, must be either "4km" or
"800m". Should only be specified for \code{temp_period} of "normals".}
}
\value{
data.frame containing metadata for all specified dates.
}
\description{
Retrieves prism metadata for a given type, temporal period, and dates. The
metadata is retrieved from the .info.txt files.
}
\details{
The metadata includes the following variables from the .info.txt file for
daily, monthly, and annual data:
\itemize{
\item PRISM_DATASET_FILENAME
\item PRISM_DATASET_CREATE_DATE
\item PRISM_DATASET_TYPE
\item PRISM_DATASET_VERSION
\item PRISM_CODE_VERSION
\item PRISM_DATASET_REMARKS
}

Additionally, two local variables are added identifying where the file is
located on the local system:
\itemize{
\item file_path
\item folder_path
}

The annual and monthly normals data includes different keys in
the .info.txt, so they are renamed to be the same as those found in the
other temporal data. The keys/variables are renamed as follows:
\itemize{
\item PRISM_FILENAME --> PRISM_DATASET_FILENAME
\item PRISM_CREATE_DATE --> PRISM_DATASET_CREATE_DATE
\item PRISM_DATASET --> PRISM_DATASET_TYPE
\item PRISM_VERSION --> PRISM_CODE_VERSION
\item PRISM_REMARKS --> PRISM_DATASET_REMARKS
}

Additionally, the normals does not include PRISM_DATASET_VERSION, so that
variable is added with \code{NA} values.
}
