% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM_main.R
\name{PRISM_main}
\alias{PRISM_main}
\title{Main pipeline of PRISM}
\usage{
PRISM_main(
  ListofTraits,
  ParametersTable,
  Index,
  sourceGWAS = getwd(),
  NbCores = 1,
  gzip = F,
  pU = 1e-05,
  ThreshSelectionPvalues = 5e-08/length(ListofTraits),
  keepIntermediateFiles = F,
  Minimum_MAF = 0.05,
  skipPairwise = F,
  labelGWASsig = F
)
}
\arguments{
\item{ListofTraits}{The list of traits to be processed by PRISM}

\item{ParametersTable}{The parameters table for each pair of traits, obtained
from LHC-MR}

\item{Index}{A dataframe with 2 columns, that contain respectively the name of
the variants and their LDscores.}

\item{sourceGWAS}{The folder where GWAS summary statistics files are held.}

\item{NbCores}{The number of cores to use, passed to `mclapply()`. Greatly
reduces the computation time.}

\item{gzip}{TRUE if you want the example data to be gzipped, FALSE otherwise.
gzip= F will speed up computations, but will increase disk space usage (only
relevant if keepIntermediateFiles = TRUE).}

\item{pU}{Polygenicity of the confounder U, as this parameter is not obtained
from LHC-MR. 1e-05 is a robust value, we do not advise to change it.}

\item{ThreshSelectionPvalues}{P-value threshold to select PRISM top variants, if you do not want
the default Bonferroni correction.}

\item{keepIntermediateFiles}{TRUE if you want to keep the intermediates files,
which can take a lot of disk space with a high number of traits and genetic variants.
We advise to set gzip = T in this case. FALSE if you want to remove them.}

\item{Minimum_MAF}{Genetic variants with minimum allele frequency below this value will be filtered out.}

\item{skipPairwise}{TRUE if you only want to execute the traitwise pipeline.}

\item{labelGWASsig}{TRUE if you want to also label variants that are not PRISM significant but are GWAS significant with the same threshold.
A "variant" and a "pval" columns are necessary in the GWAS summary statistics files.}
}
\description{
`PRISM_main()` is the main function of PRISM, that takes as input the list of traits to process,
the LDscore index, and the LHC-MR parameters table.

It will write in the Results/ folder, for each trait, the pleiotropic label
of all variants.
}
