% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_format.R
\name{prism_set_format}
\alias{prism_set_format}
\alias{prism_get_format}
\title{Set/get the PRISM grid download format}
\usage{
prism_set_format(data_format)

prism_get_format()
}
\arguments{
\item{data_format}{One of 'geotiff', 'bil', or 'nc'}
}
\value{
Invisibly returns data_format if it is valid.
}
\description{
The PRISM web service provides PRISM grids in Cloud Optimized GeoTIFF
(geotiff), bil, and netCDF (nc) format.
\code{prism_set_format()} lets the user specify the format that will be used by
the package, while \code{prism_get_format()} returns the format currently being
used. GeoTIFF is the default used by the package. It is recommended that
users choose one format and stick with it. The package is not setup to
handle a PRISM data download directory that has multiple formats in it.
}
