% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_get_set_keepZip.R
\name{prism_set_keepZip}
\alias{prism_set_keepZip}
\alias{prism_get_keepZip}
\title{Set/get whether zip files are kept after downloading}
\usage{
prism_set_keepZip(keepZip)

prism_get_keepZip()
}
\arguments{
\item{keepZip}{Logical. Keep if \code{TRUE}; delete if \code{FALSE}.}
}
\value{
\code{prism_get_keepZip()} returns the current value of \code{keepZip}.
\code{prism_set_keepZip()} invisibly returns the value if it is successfully set.
}
\description{
The PRISM data are downloaded as zip files and then unzipped. All of the
[get_prism_\emph{()\verb{](get_prism_data) functions have a logical argument that  controls this for each download. These functions let the user set/change the  default behavior rather than change it in every call to }get_prism_}()\verb{. The package default is }TRUE` to maintain previous, default behavior.
}
