% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_subset_folders.R
\name{prism_subset_folders}
\alias{prism_subset_folders}
\title{Subsets PRISM folders on the disk}
\usage{
prism_subset_folders(type, temp_period, years = NULL, mon = NULL,
  minDate = NULL, maxDate = NULL, dates = NULL, resolution = NULL)
}
\arguments{
\item{type}{The type of data you want to subset. Must be "ppt", "tmean",
"tmin", "tmax", "tdmean", "vpdmin", or "vpdmax".}

\item{temp_period}{The temporal period to subset. Must be "annual",
"monthly", "daily", "monthly normals", or "annual normals".}

\item{years}{Valid numeric year, or vector of years.}

\item{mon}{Valid numeric month, or vector of months.}

\item{minDate}{Date to start subsetting daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{maxDate}{Date to end subsetting daily data.  Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{dates}{A vector of daily dates to subset. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{resolution}{The spatial resolution of the data, must be either "4km" or
"800m". Should only be specified for \code{temp_period} of "normals".}
}
\description{
\code{prism_subset_folders()} subsets the PRISM folders stored on disk by type,
temporal period, and date. It looks through all of the PRISM data that have
been downloaded in the "prism.path" and returns the subset based on
\code{type}, \code{temp_period}, and specified dates.
}
\details{
\code{temp_period} must be specifed so the function can distinguish between
wanting annual data or wanting monthly data for a specified year. For example
\code{prism_subset_folders("tmean", "annual", years = 2012)} would provide only
one folder: the annual tmean folder. However,
\code{prism_subset_folders("tmean", "monthly", years = 2012)} would provide 12
folders: each monthly tmean folder for 2012.

\code{temp_period}, \code{years}, and \code{mon} can be combined in various different ways
to obtain different groupings of data. \code{years}, \code{mon}, and the daily
specifiers (\code{minDate}/\code{maxDate} or \code{dates}) are optional. Not specifying any
of those would result in getting all annual, monthly, or daily data.

\code{minDate}/\code{maxDate} or \code{dates} should only be specified for a \code{temp_period}
of "daily". Additionally, only \code{dates}, or \code{minDate} and \code{maxDate}, should be
specified, but all three should not be specified.
}
