% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_webservice.R
\name{prism_webservice}
\alias{prism_webservice}
\title{Download PRISM via webservice}
\usage{
prism_webservice(uri, keepZip = FALSE, returnName = FALSE, pre81_months = NULL)
}
\arguments{
\item{uri}{a valid PRISM webservice URI}

\item{keepZip}{TRUE or FALSE, keep zip files once they have been unzipped}

\item{returnName}{TRUE or FALSE, if TRUE the name of the file that was
downloaded is returned}

\item{pre81_months}{Numeric vector of months that will be downloaded, if
downloading data before 1981. This is so that the existance of the data can
be correctly checked, as the file includes all monthly data for a given
year.}
}
\description{
This is the workhorse function that will access the web service to download
files.
}
\examples{
\dontrun{
# Get the January 2001 mean temperature
prism_webservice(
  "http://services.nacse.org/prism/data/public/4km/tmean/200001", 
  keepZip = FALSE
)
}

}
