% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prismpredict.R
\name{prismpredict}
\alias{prismpredict}
\title{Partially Recursively Induced Structured Moderation}
\usage{
prismpredict(l, newdat, hvar = NULL)
}
\arguments{
\item{l}{An rpart output for prism or hprism tree.}

\item{newdat}{A data frame in which to look for variables with which to predict.}

\item{hvar}{The hierarchical level variable for hprism tree, which is specified within double quotes. If unspecified, prism tree is fitted instead. See Example for detail.}
}
\description{
Generate predicted values for new dataset from the result of prism or hprism object.
}
\details{
Models are built on the training set and then test observations were fed down each tree based on their \emph{x} and \emph{w} values, and a predicted value of the log survival time was estimated from the root nodes that each test observation fell into.
}
\examples{
\dontrun{
prd1<-prismpredict(l.prism,newdat=Leuk)
prd2<-prismpredict(l.hprism,newdat=Leuk,hvar="tpi")
prd3<-prismpredict(l.hprism1,newdat=Leuk,hvar="tpi.m")
}
}
