% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process_zip}
\alias{process_zip}
\title{Process pre 1980 files}
\usage{
process_zip(file, name)
}
\arguments{
\item{file}{the name of the file, should include "all", that is unzipped}

\item{name}{a vector of names of files that you want to save.}
}
\description{
Files that come prior to 1980 come in one huge zip.  This will cause them to mimic all post 1980 downloads
}
\details{
This should match all other files post 1980
}

