% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Multivariate Normal Random Variable Generator}
\usage{
rmvnorm(n, mu, Sigma)
}
\arguments{
\item{n}{An integer specifying the number of samples to generate.}

\item{mu}{A numeric vector specifying the mean of the distribution.}

\item{Sigma}{A numeric matrix specifying the covariance matrix of the distribution.}
}
\value{
A numeric matrix of random samples, where each column represents a sample.
}
\description{
This function generates random samples from a multivariate normal distribution.
}
\examples{
# Example usage:
mu <- c(0, 0)
Sigma <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
samples <- rmvnorm(100, mu, Sigma)
}
