% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run_analysis}
\alias{run_analysis}
\title{Run Analysis on All Pairwise Taxa}
\usage{
run_analysis(Y, alpha = rep(0, nrow(Y)), rhobound = 0.8, S = 1000)
}
\arguments{
\item{Y}{A matrix of data with observations in columns and variables in rows.}

\item{alpha}{A numeric vector of priors for the Dirichlet distribution. Defaults to a vector of zeros.}

\item{rhobound}{A numeric value specifying the bound for the \code{rho1} and \code{rho2} parameters. Defaults to 0.8.}

\item{S}{An integer specifying the number of bootstrap samples. Defaults to 1000.}
}
\value{
A data frame containing the results of the analysis including estimated 95% confidence intervals, minimum and maximum values for estimated covariance, and finite sample covariances.
}
\description{
This function runs a bootstrapped analysis on the input data matrix \code{Y}.
}
\examples{
# Example usage (You could also use the simulation function provided to generate sample data):
set.seed(123)
Y <- matrix(rnorm(1000), nrow = 10)
results <- run_analysis(Y)
}
