% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{safe_filename}
\alias{safe_filename}
\title{Create Safe File Name}
\usage{
safe_filename(x, replacement = "_")
}
\arguments{
\item{x}{Character string}

\item{replacement}{Replacement for invalid characters (default: "_")}
}
\value{
Safe file name
}
\description{
Converts string to safe file name
}
\examples{
safe_filename("My Analysis: Results (2024)")
}
