% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{save_session_info}
\alias{save_session_info}
\title{Save Session Info to File}
\usage{
save_session_info(file = "PRISM_session_info.txt")
}
\arguments{
\item{file}{Output file path}
}
\description{
Saves session information for reproducibility
}
