% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_prism_option}
\alias{set_prism_option}
\title{Set PRISM Options}
\usage{
set_prism_option(...)
}
\arguments{
\item{...}{Named arguments for options to set}
}
\value{
Invisible list of previous option values
}
\description{
Sets PRISM package options
}
\examples{
set_prism_option(verbose = FALSE, max_mismatches = 3)
}
