% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-drivers.R
\name{simulate_hotspots}
\alias{simulate_hotspots}
\title{simulate hotspots with motifs}
\usage{
simulate_hotspots(
  fracWithMotif = 0.4,
  nMotifs = 10000,
  seq = 1 + c(0, 3, 0, 1, 0, 2, 1, 1, 3, 3)
)
}
\arguments{
\item{fracWithMotif}{what fraction contain motifs}

\item{nMotifs}{how many motifs}

\item{seq}{most likely characters in sequence}
}
\description{
simulate hotspots with motifs
}
