% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{smart_read}
\alias{smart_read}
\title{Smart Read Table}
\usage{
smart_read(file, ...)
}
\arguments{
\item{file}{Path to file}

\item{...}{Additional arguments passed to read function}
}
\value{
Data frame
}
\description{
Automatically detects file format and reads appropriately
}
