% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{smart_write}
\alias{smart_write}
\title{Smart Write Table}
\usage{
smart_write(data, file, ...)
}
\arguments{
\item{data}{Data to write}

\item{file}{Output file path}

\item{...}{Additional arguments}
}
\description{
Automatically writes in appropriate format based on extension
}
