% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{standardize_chr_names}
\alias{standardize_chr_names}
\title{Standardize Chromosome Names}
\usage{
standardize_chr_names(chr_names, style = "UCSC")
}
\arguments{
\item{chr_names}{Vector of chromosome names}

\item{style}{Style to use: "UCSC" (chr1), "NCBI" (1), "Ensembl" (1)}
}
\value{
Standardized chromosome names
}
\description{
Converts chromosome names to standard format
}
\examples{
standardize_chr_names(c("1", "2", "X"), style = "UCSC")
}
