% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{subset_prism_folders}
\alias{subset_prism_folders}
\title{Subsets prism folders on the disk by type and date}
\usage{
subset_prism_folders(type, dates)
}
\arguments{
\item{type}{The type of data you want to subset. Should be tmax, tmin, tmean, 
ppt, vpdmin, or vpdmax}

\item{dates}{A vector of the dates you wish to subset as a string}
}
\description{
Looks through all of the PRISM data that is donwloaded in your \code{prism.path}
and returns the subset based on \code{type} and \code{dates}.
}
