% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-update.R
\name{summarize_probe_annotations}
\alias{summarize_probe_annotations}
\title{Summarize Probe Annotation Results}
\usage{
summarize_probe_annotations(annotations)
}
\arguments{
\item{annotations}{Data frame from update_gene_annotations}
}
\value{
List with summary statistics
}
\description{
Creates summary statistics for annotation results
}
