% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{summary_expr}
\alias{summary_expr}
\title{Quick Summary of Expression Matrix}
\usage{
summary_expr(expr_matrix)
}
\arguments{
\item{expr_matrix}{Expression matrix}
}
\value{
List with summary statistics
}
\description{
Provides quick overview of expression data
}
\examples{
\dontrun{
  summary_expr(my_matrix)
}
}
