% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{truncate_string}
\alias{truncate_string}
\title{Truncate String}
\usage{
truncate_string(x, max_length = 50)
}
\arguments{
\item{x}{Character vector}

\item{max_length}{Maximum length (default: 50)}
}
\value{
Truncated strings
}
\description{
Truncates long strings with ellipsis
}
