% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{validate_granges}
\alias{validate_granges}
\title{Validate GRanges Object}
\usage{
validate_granges(granges, require_probe_id = TRUE)
}
\arguments{
\item{granges}{GRanges object to validate}

\item{require_probe_id}{Require probe_id metadata column (default: TRUE)}
}
\value{
TRUE if valid, stops with error otherwise
}
\description{
Checks if GRanges object is properly formatted for PRISM
}
