% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlils.R
\name{write_log}
\alias{write_log}
\title{Write to Log File}
\usage{
write_log(log_file, ...)
}
\arguments{
\item{log_file}{Path to log file}

\item{...}{Message components}
}
\description{
Appends message to log file
}
