% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism.R
\name{.prism_filter_data}
\alias{.prism_filter_data}
\title{Filter counts (and paired scale data) to a valid PRISM input}
\usage{
.prism_filter_data(counts, scale, prevalence, verbose)
}
\description{
Filter counts (and paired scale data) to a valid PRISM input
}
\keyword{internal}
