% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract_version}
\alias{extract_version}
\title{Get the resolution text string}
\usage{
extract_version(type, temporal, yr)
}
\arguments{
\item{type}{the type of data you're downloading, should be tmax, tmin etc...}

\item{temporal}{The temporal resolution of the data, monthly, daily, etc...}

\item{yr}{the year of data that's being requested, in numeric form}
}
\description{
To account for the ever changing name structure, here we will scrape the HTTP directory listing and grab it instead of relying on hard coded strings that need changing
}

