% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gen_dates}
\alias{gen_dates}
\title{Processes dates as this appears many times.}
\usage{
gen_dates(minDate, maxDate, dates)
}
\arguments{
\item{minDate}{a valid iso-8601 (e.g. YYYY-MM-DD) date to start to download data. 
May be provided as either a character or \code{\link[base]{Date}} class.}

\item{maxDate}{a valid iso-8601 (e.g. YYYY-MM-DD) date to end downloading data.
May be provided as either a character or \code{\link[base]{Date}} class.}

\item{dates}{a vector of iso-8601 formatted dates to download data for, can 
also be a single date. May be provided as either a character or 
\code{\link[base]{Date}} class.}
}
\value{
Vector of dates
}
