% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_filenames}
\alias{get_filenames}
\title{List all the files in an FTP directory}
\usage{
get_filenames(type, freq, yr)
}
\arguments{
\item{type}{the}

\item{freq}{\code{character} for the frequency. One of
"daily" or "monthly".}
}
\description{
Utility function to download all filenames
for a given directory.
}

