% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prism_station_md.R
\name{get_prism_station_md}
\alias{get_prism_station_md}
\title{Extract prism metadata}
\usage{
get_prism_station_md(type, minDate = NULL, maxDate = NULL, dates = NULL)
}
\arguments{
\item{type}{The type of data to download. Must be "ppt", "tmean", "tmin",
"tmax", "tdmean", "vpdmin", or "vpdmax". Note that \code{tmean ==  mean(tmin, tmax)}.}

\item{minDate}{Date to start downloading daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{maxDate}{Date to end downloading daily data.  Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}

\item{dates}{A vector of dates to download daily data. Must be specified in
a valid iso-8601 (e.g. YYYY-MM-DD) format. May be provided as either a
character or \link[base:Date]{base::Date} class.}
}
\value{
A \code{tbl_df} containing metadata on the stations used for the specified
day and variable.
}
\description{
\code{get_prism_station_md()} extracts prism metadata on the stations used to
generate a particular day and variable. This only extracts metadata from
daily PRISM data. \strong{The daily data must already be downloaded and available
in the prism download folder.}
}
\examples{
\dontrun{
# download and then get meta data for January 1, 2010 precipitation
get_prism_dailys("ppt", dates = "2010-01-01")
get_prism_station_md("ppt", dates = "2010-01-01")
}
  
}
\seealso{
\code{\link[=get_prism_dailys]{get_prism_dailys()}}
}
