% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairwise_pipeline.R
\name{Pairwise_pipeline}
\alias{Pairwise_pipeline}
\title{Pairwise pipeline of PleioVar}
\usage{
Pairwise_pipeline(ListofTraits, ParametersTable, Index, NbCores, gzip, pU)
}
\arguments{
\item{ListofTraits}{The list of traits to be processed.}

\item{ParametersTable}{The parameters table for each pair of traits.}

\item{Index}{A dataframe with all variants and their LDscores.}

\item{NbCores}{The number of cores to use.}

\item{gzip}{TRUE to gzip created files, FALSE otherwise.}

\item{pU}{Polygenicity of the confounder U.}
}
\description{
`Pairwise_pipeline()` handles the pairwise process of PleioVar. It takes as input
the data and outputs in the Traitwise/ folder the likelihoods and scores of
each variant to belong to the different models, for each pair of traits.


All parameters are passed from `PleioVar_main()` and are more detailed over there.
}
