% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prism_md.R
\name{pr_parse}
\alias{pr_parse}
\title{name parse}
\usage{
pr_parse(p)
}
\arguments{
\item{p}{a prism file directory}
}
\value{
a properly parsed string of human readable names
}
\description{
parse the directory name into relevant metadata
}

