% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_md.R
\name{pr_parse}
\alias{pr_parse}
\title{name parse}
\usage{
pr_parse(p, returnDate = FALSE)
}
\arguments{
\item{p}{a prism file directory}

\item{returnDate}{TRUE or FALSE?  If TRUE, an ISO date is returned.  By default years will come back with YYYY-01-01 and months as YYYY-MM-01}
}
\value{
a properly parsed string of human readable names
}
\description{
parse the directory name into relevant metadata
}

