% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{print.prism_result}
\alias{print.prism_result}
\title{Print a PRISM result}
\usage{
\method{print}{prism_result}(x, ...)
}
\arguments{
\item{x}{A `prism_result`.}

\item{...}{Unused.}
}
\description{
Print a PRISM result
}
