% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism.R
\docType{package}
\name{prism-package}
\alias{prism}
\alias{prism-package}
\title{prism: Access Data from the Oregon State Prism Climate Project}
\description{
Allows users to access the Oregon State Prism climate data
    (<https://www.prism.oregonstate.edu/>). Using the web service API data
    can easily downloaded in bulk and loaded into R for spatial analysis.
    Some user friendly visualizations are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/prism/}
  \item \url{https://github.com/ropensci/prism}
  \item Report bugs at \url{https://github.com/ropensci/prism/issues}
}

}
\author{
\strong{Maintainer}: Alan Butler \email{rabutler@usbr.gov} [contributor]

Authors:
\itemize{
  \item Hart Edmund \email{Edmund.m.hart@gmail.com} [conceptor]
  \item Kendon Bell \email{kmb56@berkeley.edu}
}

}
\keyword{internal}
