% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prism_check}
\alias{prism_check}
\title{Helper function to check if files already exist}
\usage{
prism_check(prismfiles)
}
\arguments{
\item{prismfiles}{a list of full paths for prism files}
}
\value{
a character vector of file names that already exist
}
\description{
check if files exist
}

