% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prism_check}
\alias{prism_check}
\title{Helper function to check if files already exist}
\usage{
prism_check(prismfiles, lgl = FALSE)
}
\arguments{
\item{prismfiles}{a list of full paths for prism files}

\item{lgl}{\code{TRUE} returns a logical vector indicating those
not yet downloaded; \code{FALSE}
returns the file names that are not yet downloaded.}
}
\value{
a character vector of file names that are not yet downloaded
or a logical vector indication those not yet downloaded..
}
\description{
check if files exist
}

