% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_image.R
\name{prism_image}
\alias{prism_image}
\title{Quick spatial image of prism data}
\usage{
prism_image(prismfile, col = "heat")
}
\arguments{
\item{prismfile}{the name of a single file to be plotted, this is most easily
gotten through \code{\link[=ls_prism_data]{ls_prism_data()}} or \code{\link[=prism_data_subset]{prism_data_subset()}}. It should be the
folder name, not the .bil file.}

\item{col}{the color pattern to use.  The default is heat, the other valid
option is "redblue".}
}
\value{
Invisibly returns \code{gg} object of the image.
}
\description{
\code{prism_image()} quickly makes a spatial image plot of the specified prism
data (single variable and time step.). It is meant for rapid visualization,
but more detailed plots will require other methods.
}
\examples{
\dontrun{
get_prism_dailys(
  type = "tmean", 
  minDate = "2013-06-01", 
  maxDate = "2013-06-14", 
  keepZip = FALSE
)

# get June 5th 
pd <- prism_data_subset("tmean", "daily", dates = "2013-06-05")

# and plot it
prism_image(pd)
}

}
\seealso{
\code{\link[=ls_prism_data]{ls_prism_data()}}, \code{\link[=prism_data_subset]{prism_data_subset()}}, \code{\link[ggplot2:geom_raster]{ggplot2::geom_raster()}}
}
