% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_md.R
\name{prism_md}
\alias{prism_md}
\title{Extract select prism metadata}
\usage{
prism_md(f, returnDate = FALSE)
}
\arguments{
\item{f}{a simple directory name}

\item{returnDate}{TRUE or FALSE?  If TRUE, an ISO date is returned.  By default years will come back with YYYY-01-01 and months as YYYY-MM-01}
}
\value{
a character vector of metadata.
}
\description{
used to extract some prism metadata used in other fxns
}
