% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prism_md.R
\name{prism_md}
\alias{prism_md}
\title{Extract select prism metadata}
\usage{
prism_md(f)
}
\arguments{
\item{f}{a simple directory name}
}
\value{
a character vector of metadata.
}
\description{
used to extract some prism metadata used in other fxns
}

