% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_set_dl_dir.R
\name{prism_set_dl_dir}
\alias{prism_set_dl_dir}
\title{Set prism download directory}
\usage{
prism_set_dl_dir(path, create = TRUE)
}
\arguments{
\item{path}{The path that prism data will be unzipped into.}

\item{create}{Boolean that determines if the \code{path} will be created if it
does not already exist.}
}
\description{
Sets the directory that downloaded prism data will be unzipped into.
}
