% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prism_stack.R
\name{prism_stack}
\alias{prism_stack}
\title{Stack Prism files}
\usage{
prism_stack(prismfile)
}
\arguments{
\item{prismfile}{a vector of file names returned by ls_prism_data()}
}
\description{
Create a stack of prism files
}
\examples{
\dontrun{
get_prism_dailys(type="tmean", minDate = "2013-06-01", maxDate = "2013-06-14", keepZip=FALSE)
mystack <- prism_stack(ls_prism_data()[1:14])
}
}

