% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prism_webservice.R
\name{prism_webservice}
\alias{prism_webservice}
\title{Download PRISM via webservice}
\usage{
prism_webservice(uri, keepZip = FALSE, returnName = FALSE)
}
\arguments{
\item{uri}{a valid PRISM webservice URI}

\item{keepZip}{TRUE or FALSE, keep zip files once they have been unzipped}

\item{returnName}{TRUE or FALSE, if TRUE the name of the file that was downoaded is returned}
}
\description{
This is the workhorse function that will access the web service to download files
}
\examples{
\dontrun{
### Get all the data for January from 1990 to 2000
get_prism_annual(type="tmean", year = 1990:2000, keepZip=FALSE)
}
}

