% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process_zip}
\alias{process_zip}
\title{Process pre 1980 files}
\usage{
process_zip(pfile, name)
}
\arguments{
\item{pfile}{the name of the file, should include "all", that is unzipped}

\item{name}{a vector of names of files that you want to save.}
}
\description{
Files that come prior to 1980 come in one huge zip.  This will cause them to mimic all post 1980 downloads
}
\details{
This should match all other files post 1980
}
\examples{
\dontrun{
process_zip('PRISM_tmean_stable_4kmM2_1980_all_bil','PRISM_tmean_stable_4kmM2_198001_bil')
process_zip('PRISM_tmean_stable_4kmM2_1980_all_bil',c('PRISM_tmean_stable_4kmM2_198001_bil','PRISM_tmean_stable_4kmM2_198002_bil'))
}
}

