% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaplots.R
\name{sigmaplot}
\alias{sigmaplot}
\title{Plot Sigma Values Against Parameters}
\usage{
sigmaplot(
  all_inner_results,
  bg = "white",
  filename = NULL,
  save = NULL,
  individual = FALSE,
  dir_path = "./plots/",
  plot_type = "line"
)
}
\arguments{
\item{all_inner_results}{A data frame containing sigma values and corresponding parameters.}

\item{bg}{A character string indicating the background color of the plot. Options are "white" (default) or "transparent".}

\item{filename}{A character string specifying an alternate name of the file. Defaults to \code{NULL} which saves as parameter combination names.}

\item{save}{A character string specifying the format to save the plots (\code{NULL}, \code{"jpg"}, \code{"png"}, \code{"svg"}). Defaults to \code{NULL}.}

\item{individual}{A boolean type specifying whether you would like to save individual plots. Defaults to \code{FALSE}.}

\item{dir_path}{A character string indicating the directory to store the plot. Default is \code{"./plots/"} which creates the plots directory in the current directory.}

\item{plot_type}{A character string specifying the type of plot ("point" or "line"). Defaults to "line". "line" fits a quadratic line of best fit to the data points. "point" plots the individual points.}
}
\value{
Plots visualizing the impact of the parameters on sigma.
}
\description{
This function creates plots to visualize the impact of parameters \code{rho1}, \code{rho2}, and \code{x} on sigma values.
It identifies and highlights the parameter points that minimize and maximize sigma for each comparison.
Optionally, it can save the plots in \code{jpg}, \code{png}, or \code{svg} format with 300 dpi resolution.
}
\examples{
# Example usage:
# Assuming 'results' is the output from estimate_covariance function
# results <- estimate_covariance(Y)
# all_inner_results <- results$all_inner_results # OR 
# all_inner_results <- results$all_inner_results_moment
# sigmaplot(all_inner_results)  # Without saving
# sigmaplot(all_inner_results, save = "png", filename = "sampledataset", individual = FALSE, plot_type = "line")  # Save as PNG with best fit lines
}
