% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Traitwise_pipeline.R
\name{Traitwise_pipeline}
\alias{Traitwise_pipeline}
\title{Traitwise pipeline of PleioVar}
\usage{
Traitwise_pipeline(
  ListofTraits,
  ParametersTable,
  Index,
  sourceGWAS,
  NbCores,
  gzip,
  pU,
  ThreshSelectionPvalues,
  labelGWASsig
)
}
\arguments{
\item{ListofTraits}{The list of traits to be processed.}

\item{ParametersTable}{The parameters table for each pair of traits.}

\item{Index}{A dataframe with all variants and their LDscores.}

\item{sourceGWAS}{The folder where GWAS summary statistics files are held.}

\item{NbCores}{The number of cores to use.}

\item{gzip}{TRUE to gzip created files, FALSE otherwise.}

\item{pU}{Polygenicity of the confounder U.}

\item{ThreshSelectionPvalues}{P-value threshold to select PleioVar top variants.}

\item{labelGWASsig}{TRUE if you want to label variants that are not PleioVar significant but are GWAS significant.}
}
\description{
`Traitwise_pipeline()` handles the traitwise process of PleioVar. It takes as input
the likelihoods and scores obtained from `Pairwise_pipeline()`. It outputs the syntheses
and the PleioVar p-values by traits in the Traitwise/ folder.

Final results (PleioVar p-values and pleiotropic labels, for all variants) are
written for each trait in the Results/ folder.

All parameters are passed from `PleioVar_main()` and are more detailed over there.
}
